agents/web-researcher/requirements.txt
openai>=1.12.0
streamlit>=1.30.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
